fixup commit for branch 'MOZILLA_0_9_7_BRANCH'
--HG--
branch : MOZILLA_0_9_7_BRANCH

+++ dbm/tests/lots.c
@@ -1,24 +1,40 @@
-/* -*- Mode: C; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 2 -*-
+/* -*- Mode: C; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
+/* ***** BEGIN LICENSE BLOCK *****
+ * Version: NPL 1.1/GPL 2.0/LGPL 2.1
  *
- * The contents of this file are subject to the Netscape Public
- * License Version 1.1 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of
- * the License at http://www.mozilla.org/NPL/
+ * The contents of this file are subject to the Netscape Public License
+ * Version 1.1 (the "License"); you may not use this file except in
+ * compliance with the License. You may obtain a copy of the License at
+ * http://www.mozilla.org/NPL/
  *
- * Software distributed under the License is distributed on an "AS
- * IS" basis, WITHOUT WARRANTY OF ANY KIND, either express or
- * implied. See the License for the specific language governing
- * rights and limitations under the License.
+ * Software distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
+ * for the specific language governing rights and limitations under the
+ * License.
  *
  * The Original Code is mozilla.org code.
  *
- * The Initial Developer of the Original Code is Netscape
- * Communications Corporation.  Portions created by Netscape are
- * Copyright (C) 1998 Netscape Communications Corporation. All
- * Rights Reserved.
+ * The Initial Developer of the Original Code is 
+ * Netscape Communications Corporation.
+ * Portions created by the Initial Developer are Copyright (C) 1998
+ * the Initial Developer. All Rights Reserved.
  *
- * Contributor(s): 
- */
+ * Contributor(s):
+ *
+ *
+ * Alternatively, the contents of this file may be used under the terms of
+ * either the GNU General Public License Version 2 or later (the "GPL"), or
+ * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
+ * in which case the provisions of the GPL or the LGPL are applicable instead
+ * of those above. If you wish to allow use of your version of this file only
+ * under the terms of either the GPL or the LGPL, and not to allow others to
+ * use your version of this file under the terms of the NPL, indicate your
+ * decision by deleting the provisions above and replace them with the notice
+ * and other provisions required by the GPL or the LGPL. If you do not delete
+ * the provisions above, a recipient may use your version of this file under
+ * the terms of any one of the NPL, the GPL or the LGPL.
+ *
+ * ***** END LICENSE BLOCK ***** */
 
 /* use sequental numbers printed to strings
  * to store lots and lots of entries in the
@@ -123,7 +139,7 @@ DBT * MakeLargeKey(int32 num)
 	static DBT rv;
 	static char *string_rv=0;
 	int rep_char;
-	int32 size;
+	size_t size;
 
 	if(string_rv)
 		free(string_rv);
@@ -138,9 +154,9 @@ DBT * MakeLargeKey(int32 num)
 
 	/* malloc a string low_bits wide */
 	size = low_bits*sizeof(char);
-	string_rv = (char *)malloc((size_t)size);
+	string_rv = (char *)malloc(size);
 
-	memset(string_rv, rep_char, (size_t)size);
+	memset(string_rv, rep_char, size);
 
 	rv.data = string_rv;
 	rv.size = size;
@@ -207,7 +223,7 @@ int
 VerifyData(DBT *data, int32 num, key_type_enum key_type)
 {
 	int32 count, compare_num;
-	uint32 size;
+	size_t size;
 	int32 *int32_array;
 
 	/* The first int32 is count 
@@ -311,7 +327,7 @@ GenData(int32 num)
 	int32 n;
 	static DBT *data=0;
 	int32 *int32_array;
-	int32 size;
+	size_t size;
 
 	if(!data)
 	  {
@@ -330,7 +346,7 @@ GenData(int32 num)
 
 	
 	size = sizeof(int32)*(n+1);
-	int32_array = (int32 *) malloc((size_t)size);
+	int32_array = (int32 *) malloc(size);
 
 	memcpy(&int32_array[0], &n, sizeof(int32));
